Force instance name to be defined in vagrant deploy
