+++ deployability/modules/allocation/aws/provider.py
@@ -354,22 +354,3 @@ class AWSProvider(Provider):
                 logger.warning(f"{message}")
         else:
             logger.info(f"Dedicated host released: {host_identifier}")
-
-    @staticmethod
-    def generate_repository_name(repository: str) -> str:
-        """
-        Generate a repository name for the instance.
-
-        Args:
-            repository (str): Repository name.
-
-        Returns:
-            str: Repository name for the instance.
-        """
-        matches = re.findall(r'(\w+)', repository)
-        if len(matches) == 3:
-            return ''.join([c[0] for c in matches])
-        elif len(matches) == 2:
-            return matches[1]
-        else:
-            return repository
